updated easylogging++
