+++ CMakeLists.txt
@@ -12,7 +12,7 @@ option (old_toolchain "Is old toolchain" OFF)
 
 set (RESIDUE_MAJOR "2")
 set (RESIDUE_MINOR "3")
-set (RESIDUE_PATCH "3")
+set (RESIDUE_PATCH "4")
 set (RESIDUE_VERSION "${RESIDUE_MAJOR}.${RESIDUE_MINOR}.${RESIDUE_PATCH}")
 set (RESIDUE_NAME "Residue")
